[FIX] mass_mailing: make better link dialog for mass mailing (2)
Commit [1] had to be adapted before being forward-ported.

[1]: https://github.com/odoo/odoo/commit/70da865a042dd6ddb012c30d97258692f276e2a0
